Add E2E test for browser extension credential create flow (#643)
