+++ browser-extensions/chrome/src/shared/__tests__/AppInfo.test.ts
@@ -30,6 +30,7 @@ describe('AppInfo', () => {
 
     it('should reject lower versions', () => {
       expect(AppInfo.versionGreaterThanOrEqualTo('1.0.0', '1.0.1')).toBe(false);
+      expect(AppInfo.versionGreaterThanOrEqualTo('1.0.0', '1.4.1')).toBe(false);
       expect(AppInfo.versionGreaterThanOrEqualTo('1.4.0', '1.5.0')).toBe(false);
       expect(AppInfo.versionGreaterThanOrEqualTo('1.9.9', '2.0.0')).toBe(false);
     });
